training/yolo-plant-detection/requirements.txt
numpy>=2.1.0
opencv-python-headless>=4.10.0
pandas>=2.2.0
psycopg[binary]>=3.2.0
pyyaml>=6.0.1
scikit-learn>=1.5.0
tqdm>=4.66.0
ultralytics>=8.3.0
